messages_app/serializers.py
from rest_framework import serializers
from messages_app.models import Message
from messages_app.models import User


class MessagesSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = '__all__'


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'